refactor(pkix): switch to crypto/x509's CertificateRequests
As of Go 1.3, PKCS #10 CSR support is present in the Go standard
library. This change removes the third-party pkcs10 dependency, and
switches the pkix CSR handling code to use the standard library.

The checkSignature function added to csr.go comes from CFSSL
(https://github.com/cloudflare/cfssl), which is covered under the
3-clause BSD license.